This is MEDYANv3.0 for comparison purposes

+++ src/RateChanger.h
@@ -1,9 +1,9 @@
 
 //------------------------------------------------------------------
 //  **MEDYAN** - Simulation Package for the Mechanochemical
-//               Dynamics of Active Networks, v3.1
+//               Dynamics of Active Networks, v3.0
 //
-//  Copyright (2015-2016)  Papoian Lab, University of Maryland
+//  Copyright (2015)  Papoian Lab, University of Maryland
 //
 //                 ALL RIGHTS RESERVED
 //
@@ -68,7 +68,7 @@ public:
     /// Change the reaction rate based on an on rate, off rate,
     /// number of heads, and given force.
     virtual float changeRate(float onRate, float offRate,
-                             double numHeads, double force) = 0;
+                             int numHeads, double force) = 0;
 };
 
 #endif
